feat: map plugin (#2229)
