+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-02-10T10:21:06.817Z\n"
-"PO-Revision-Date: 2023-02-10T10:21:06.817Z\n"
+"POT-Creation-Date: 2023-02-14T12:38:49.529Z\n"
+"PO-Revision-Date: 2023-02-14T12:38:49.529Z\n"
 
 msgid "Untitled dashboard"
 msgstr "Untitled dashboard"
@@ -104,8 +104,11 @@ msgstr "There was a problem loading interpretations for this item"
 msgid "Maps with Earth Engine layers cannot be displayed when offline"
 msgstr "Maps with Earth Engine layers cannot be displayed when offline"
 
-msgid "Unable to load the plugin for this item"
-msgstr "Unable to load the plugin for this item"
+msgid "The plugin for rendering this item is not available"
+msgstr "The plugin for rendering this item is not available"
+
+msgid "Install the {{appName}} app from the App Hub"
+msgstr "Install the {{appName}} app from the App Hub"
 
 msgid "The plugin for rendering this item is not available"
 msgstr "The plugin for rendering this item is not available"
@@ -121,6 +124,9 @@ msgstr "Filters are not applied to line list dashboard items."
 
 msgid "Show without filters"
 msgstr "Show without filters"
+
+msgid "Unable to load the plugin for this item"
+msgstr "Unable to load the plugin for this item"
 
 msgid "There was an error loading data for this item"
 msgstr "There was an error loading data for this item"
