+++ src/components/Item/VisualizationItem/Visualization/plugin.js
@@ -12,20 +12,17 @@ import { loadExternalScript } from './loadExternalScript.js'
 
 //external plugins
 const itemTypeToGlobalVariable = {
-    [MAP]: 'mapPlugin',
     [EVENT_REPORT]: 'eventReportPlugin',
     [EVENT_CHART]: 'eventChartPlugin',
 }
 
 const itemTypeToScriptPath = {
-    [MAP]: '/dhis-web-maps/map.js',
     [EVENT_REPORT]: '/dhis-web-event-reports/eventreport.js',
     [EVENT_CHART]: '/dhis-web-event-visualizer/eventchart.js',
 }
 
 const hasIntegratedPlugin = (type) =>
-    // TODO add MAP here when its new plugin is available
-    [CHART, REPORT_TABLE, VISUALIZATION].includes(type)
+    [CHART, REPORT_TABLE, VISUALIZATION, MAP].includes(type)
 
 export const getPluginLaunchUrl = (type, d2) => {
     // 1. lookup in api/apps for the "manually installed" app, this can be a new version for a core (bundled) app
@@ -46,6 +43,9 @@ export const getPluginLaunchUrl = (type, d2) => {
             case REPORT_TABLE:
             case VISUALIZATION: {
                 return `${baseUrl}/dhis-web-data-visualizer/plugin.html`
+            }
+            case MAP: {
+                return `${baseUrl}/dhis-web-maps/plugin.html`
             }
         }
     }
@@ -93,7 +93,7 @@ const fetchPlugin = async (type, baseUrl) => {
 export const pluginIsAvailable = (type, d2) =>
     hasIntegratedPlugin(type) ||
     Boolean(getPluginLaunchUrl(type, d2)) ||
-    itemTypeToGlobalVariable[type]
+    Boolean(itemTypeToGlobalVariable[type])
 
 const loadPlugin = async ({ type, config, credentials, d2 }) => {
     if (!pluginIsAvailable(type, d2)) {
